requirements.txt
awpy==1.3.1
pandas==2.2.3
patool==2.4.0
